Use long instaed of int array in sha512

+++ multiplatform-crypto-libsodium-bindings/src/jvmMain/kotlin/com/ionspin/kotlin/crypto/JnaLibsodiumInterface.kt
@@ -26,7 +26,7 @@ class Hash512State : Structure() {
     override fun getFieldOrder(): List<String> = listOf("state", "count", "buf")
 
     @JvmField
-    val state = IntArray(8)
+    val state = LongArray(8)
 
     @JvmField
     var count: LongArray = LongArray(2)
